% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district_map}
\alias{district_map}
\title{district_map}
\format{An object of class \code{SpatialPolygonsDataFrame} with 74 rows and 15 columns.}
\source{
Humanitarian Data Exchange <https://data.humdata.org/dataset/somalia-administrative-boundaries>
}
\usage{
district_map
}
\description{
Administrative Level 2 (District) map of Somalia
}
\keyword{datasets}
