% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhz2015}
\alias{lhz2015}
\title{lhz2015}
\format{A SpatialPolygonsDataFrame with 19 features:
\describe{
\item{\code{COUNTRY}}{2-digit ISO country code (SO)}
\item{\code{LZNUM}}{Livelihood zone unique identifier}
\item{\code{LZCODE}}{Livelihood zone code}
\item{\code{LZNAME}}{Livelihood name}
\item{\code{CLASS}}{Livelihood zone classification}
}}
\source{
FEWS-NET
}
\usage{
lhz2015
}
\description{
Livelihood zones map of Somalia (2015)
}
\keyword{datasets}
