% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villages}
\alias{villages}
\title{villages}
\format{A data frame with 9 columns and 9782 rows:
\describe{
\item{\code{admin1pcode}}{Administrative level 1 postal code}
\item{\code{admin1id}}{Administrative level 1 unique identifier}
\item{\code{admin1name}}{Administrative level 1 name}
\item{\code{admin2id}}{Administrative level 2 unique identifier}
\item{\code{admin2name}}{Administrative level 2 name}
\item{\code{admin3id}}{Administrative level 3 unique identifier}
\item{\code{admin3name}}{Administrative level 3 name}
\item{\code{longitude}}{Longitude coordinate in GPS format}
\item{\code{latitude}}{Latitude coordinate in GPS format}
}}
\source{
UNICEF Somalia
}
\usage{
villages
}
\description{
List of villages in Somalia with corresponding GPS coordinates.
}
\keyword{datasets}
