% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_map}
\alias{region_map}
\title{region_map}
\format{An object of class \code{SpatialPolygonsDataFrame} with 18 rows and 13 columns.}
\source{
Humanitarian Data Exchange <https://data.humdata.org/dataset/somalia-administrative-boundaries>
}
\usage{
region_map
}
\description{
Administrative Level 1 (Region) map of Somalia
}
\keyword{datasets}
